refactor(try one by one every route tomorrow)
