--- api/requests/auth.request.http
@@ -1,30 +1,0 @@
-@api = http://localhost:3001/api/v1/auth
-
-### verify token
-POST {{api}}/verify HTTP/1.1
-content-type: application/json
-
-{
-    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdhcmNpYXNhYWliQGdtYWlsLmNvbSIsIm5pY2tuYW1lIjoiZ2FyY2lhc2FhaWIiLCJhY2Nlc3MiOnRydWUsImlhdCI6MTY0ODQ0MTIzOSwiZXhwIjoxNjQ4NTI3NjM5fQ.i_LJGJXc1qmy_mvc7S8JqmIsTmOYVcf38iOju9WCAmk"
-}
-
-### get signin metadata
-GET {{api}}/create HTTP/1.1
-
-### signin
-POST {{api}}/signin HTTP/1.1
-content-type: application/json
-
-{
-    "email": "garciasaaib@gmail.com",
-    "password": "garciasaaib"
-}
-
-### signup
-POST {{api}}/signup HTTP/1.1
-content-type: application/json
-
-{
-    "email": "garciasaaib@gmail.com",
-    "password": "garciasaaib1"
-}
