--- api/requests/client.request.http
@@ -1,39 +1,0 @@
-@api = http://localhost:3001/api/v1/client
-
-
-
-### get signin metadata
-GET {{api}}/ HTTP/1.1
-
-### get signin metadata
-GET {{api}}/create HTTP/1.1
-
-### login as a client
-POST {{api}}/login HTTP/1.1
-content-type: application/json
-
-{
-    "email": "garciasaaib@gmail.com",
-    "password": "garciasaaib"
-}
-
-
-### new client acount
-POST {{api}}/register HTTP/1.1
-content-type: application/json
-
-{
-    "email": "garciasaaaib@gmail.com",
-    "password": "garciasaaib",
-    "firstname": "adrian",
-    "lastname": "garcia",
-    "secondlastname": "saaib"
-}
-
-### forgot password
-POST {{api}}/forgotpassword HTTP/1.1
-content-type: application/json
-
-{
-    "email": "garciasaaib@gmail.com"
-}
