--- api/requests/employee.request.http
@@ -1,14 +1,0 @@
-
-
-@endpoint = http://localhost:3001/api/v1/employees
-### Room list with the info to check if is avaiable or not & why
-GET {{endpoint}} HTTP/1.1
-
-
-###
-GET http://localhost:3001/api/v1/employee HTTP/1.1
-
-###
-POST http://localhost:3001/api/v1/employee HTTP/1.1
-
-
