--- api/requests/info.request.http
@@ -1,11 +1,0 @@
-### vars
-# @enpoint = http://localhost:3001/api/v1/info
-
-@api = {{baseUrl}}{{port}}/api/{{version}}
-@endpoint = {{api}}/info
-
-### get client show info
-GET {{endpoint}}/client HTTP/1.1
-
-### get employee show info
-GET {{endpoint}}/employee HTTP/1.1
