+++ api/src/database/models/index.js
@@ -6,7 +6,7 @@ const Sequelize = require('sequelize');
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
 // const config = require(__dirname + '/../config/database')[env];
-const config = require(__dirname + '/../config/database.js')[env]
+const config = require(__dirname + '/../../config/database.js')[env]
 const db = {};
 
 let sequelize;
