+++ api/src/database/models/roomphoto.js
