+++ api/src/routes/auth/auth.request.http
@@ -1,0 +1,36 @@
+### LOGIN 
+POST {{api}}/login HTTP/1.1
+content-type: application/json
+
+{
+    "email": "garciasaaib@gmail.com",
+    "password": "garciasaaib"
+}
+
+
+
+### GENERATE MAIL LINK TO UPDATE PASSWORD
+POST {{api}}/forgotpassword HTTP/1.1
+content-type: application/json
+
+{
+    "email": "garciasaaib@gmail.com"
+}
+
+
+
+### SET NEW PASSWORD
+PUT {{api}}/newpassword HTTP/1.1
+content-type: application/json
+
+{
+    "password": ""
+}
+
+### SET TRUE VERIFY EMAIL
+PUT {{api}}/verifyemail/ HTTP/1.1
+content-type: application/json
+
+{
+    "password": ""
+}
