--- api/src/routes/info.router.js
@@ -1,7 +1,0 @@
-import { Router } from "express"
-import infoController from '../controllers/info.controller'
-
-const app = Router()
-  // app.get('/:type?', infoController.show)
-
-export default app
