+++ api/src/routes/room/room.request.http
@@ -1,18 +1,17 @@
 ### vars
 @endpoint = http://localhost:3001/api/v1/rooms
-# @api = {{baseUrl}}{{port}}/api/{{version}}
-# @endpoint = {{api}}/rooms
 
 
 
-### show all rooms
+
+### SHOW ALL ROOMS
 GET {{endpoint}} HTTP/1.1
 
-### get client show all
+### SHOW ALL ROOMS & FILTER BY QUERYS
 GET {{endpoint}}?page=3 HTTP/1.1
 
-### show all room types
+### SHOW ALL TYPES OF ROOMS
 GET {{endpoint}}/types HTTP/1.1
 
-### show details for one room type
+### SHOW DETAILS OF A TYPE OF ROOMS
 GET {{endpoint}}/types/3 HTTP/1.1
